organization of methods on CommunityPatchSubModule
clean up using statements
