+++ src/CommunityPatch/CommunityPatchSubModule.CopyDiagnosticsToClipboard.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Numerics;
-using System.Reflection;
 using System.Runtime;
 using System.Runtime.InteropServices;
 using System.Text;
