+++ src/CommunityPatch/CommunityPatchSubModule.Initialization.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Reflection;
 using HarmonyLib;
-using TaleWorlds.Engine;
 
 namespace CommunityPatch {
