+++ src/CommunityPatch/CommunityPatchSubModule.cs
@@ -1,13 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using HarmonyLib;
 using JetBrains.Annotations;
 using TaleWorlds.CampaignSystem;
 using TaleWorlds.Core;
-using TaleWorlds.Engine;
-using TaleWorlds.Library;
 using TaleWorlds.Localization;
 using TaleWorlds.MountAndBlade;
 using Module = TaleWorlds.MountAndBlade.Module;
@@ -17,34 +13,8 @@ namespace CommunityPatch {
   [PublicAPI]
   public partial class CommunityPatchSubModule : MBSubModuleBase {
 
-    internal static readonly Harmony Harmony = new Harmony(nameof(CommunityPatch));
-
-    internal static readonly LinkedList<Exception> RecordedFirstChanceExceptions
-      = new LinkedList<Exception>();
-
-    internal static readonly LinkedList<Exception> RecordedUnhandledExceptions
-      = new LinkedList<Exception>();
-
-    internal static readonly OptionsFile Options = new OptionsFile(nameof(CommunityPatch) + ".txt");
-
     [PublicAPI]
     internal static CampaignGameStarter CampaignGameStarter;
-
-    internal static bool DisableIntroVideo {
-      get => Options.Get<bool>(nameof(DisableIntroVideo));
-      set => Options.Set(nameof(DisableIntroVideo), value);
-    }
-
-    internal static bool RecordFirstChanceExceptions {
-      get => Options.Get<bool>(nameof(RecordFirstChanceExceptions));
-      set => Options.Set(nameof(RecordFirstChanceExceptions), value);
-    }
-
-    public override void BeginGameStart(Game game)
-      => base.BeginGameStart(game);
-
-    public override bool DoLoading(Game game)
-      => base.DoLoading(game);
 
     protected override void OnBeforeInitialModuleScreenSetAsRoot() {
       var module = Module.CurrentModule;
@@ -86,63 +56,10 @@ namespace CommunityPatch {
       ));
       base.OnSubModuleLoad();
     }
+
     private static void ShowMessage(string msg) {
       InformationManager.DisplayMessage(new InformationMessage(msg));
       Print(msg);
-    }
-
-    private void ShowModOptions() {
-      var elements = new List<InquiryElement>();
-
-      elements.Add(new InquiryElement(
-        nameof(DisableIntroVideo),
-        DisableIntroVideo ? "Enable Intro Videos" : "Disable Intro Videos",
-        null
-      ));
-
-      elements.Add(new InquiryElement(
-        nameof(RecordFirstChanceExceptions),
-        RecordFirstChanceExceptions ? "Ignore First Chance Exceptions" : "Record First Chance Exceptions",
-        null
-      ));
-      elements.Add(new InquiryElement(
-        nameof(CopyDiagnosticsToClipboard),
-        "Copy Diagnostics to Clipboard",
-        null
-      ));
-      elements.Add(new InquiryElement(
-        "IntentionallyUnhandled",
-        "Throw Unhandled Exception",
-        null
-      ));
-      InformationManager.ShowMultiSelectionInquiry(new MultiSelectionInquiryData(
-        "Mod Options",
-        "Community Patch Mod Options:",
-        elements,
-        true,
-        true,
-        "Apply",
-        "Return",
-        list => {
-          var selected = (string) list[0].Identifier;
-          switch (selected) {
-            case nameof(DisableIntroVideo):
-              DisableIntroVideo = !DisableIntroVideo;
-              ShowMessage($"Intro Videos: {(DisableIntroVideo ? "Disabled" : "Enabled")}.");
-              Options.Save();
-              break;
-            case nameof(RecordFirstChanceExceptions):
-              RecordFirstChanceExceptions = !RecordFirstChanceExceptions;
-              ShowMessage($"Record FCEs: {(RecordFirstChanceExceptions ? "Enabled" : "Disabled")}.");
-              Options.Save();
-              break;
-            case nameof(CopyDiagnosticsToClipboard):
-              CopyDiagnosticsToClipboard();
-              break;
-            default:
-              throw new NotImplementedException(selected);
-          }
-        }, null));
     }
 
     public override void OnCampaignStart(Game game, object starterObject) {
@@ -156,95 +73,6 @@ namespace CommunityPatch {
       ApplyPatches(game);
 
       base.OnGameInitializationFinished(game);
-    }
-
-    public static IDictionary<Type, IPatch> ActivePatches
-      = new Dictionary<Type, IPatch>();
-
-    private static void ApplyPatches(Game game) {
-      //ActivePatches.Clear();
-
-      foreach (var patch in Patches) {
-        try {
-          patch.Reset();
-        }
-        catch (Exception ex) {
-          Error(ex, $"Error while resetting patch: {patch.GetType().Name}");
-        }
-
-        try {
-          if (patch.IsApplicable(game)) {
-            try {
-              patch.Apply(game);
-            }
-            catch (Exception ex) {
-              Error(ex, $"Error while applying patch: {patch.GetType().Name}");
-            }
-          }
-        }
-        catch (Exception ex) {
-          Error(ex, $"Error while checking if patch is applicable: {patch.GetType().Name}");
-        }
-
-        var patchApplied = patch.Applied;
-        if (patchApplied)
-          ActivePatches[patch.GetType()] = patch;
-
-        ShowMessage($"{(patchApplied ? "Applied" : "Skipped")} Patch: {patch.GetType().Name}");
-      }
-    }
-
-    private static LinkedList<IPatch> _patches;
-
-    [PublicAPI]
-    public static readonly ApplicationVersionComparer VersionComparer
-      = new ApplicationVersionComparer();
-
-    [PublicAPI]
-    public static ApplicationVersion GameVersion
-      = ModuleInfo.GetModules()
-        .Where(x => x.IsOfficial)
-        .OrderByDescending(x => x.Version, VersionComparer)
-        .FirstOrDefault()?.Version ?? default;
-
-    public static LinkedList<IPatch> Patches {
-      get {
-        if (_patches != null)
-          return _patches;
-
-        var patchInterfaceType = typeof(IPatch);
-        _patches = new LinkedList<IPatch>();
-
-        foreach (var type in typeof(CommunityPatchSubModule).Assembly.GetTypes()) {
-          if (type.IsInterface || type.IsAbstract)
-            continue;
-          if (!patchInterfaceType.IsAssignableFrom(type))
-            continue;
-
-          try {
-            var patch = (IPatch) Activator.CreateInstance(type, true);
-            //var patch = (IPatch) FormatterServices.GetUninitializedObject(type);
-            _patches.AddLast(patch);
-          }
-          catch (TargetInvocationException tie) {
-            Error(tie.InnerException, $"Failed to create instance of patch: {type.FullName}");
-          }
-          catch (Exception ex) {
-            Error(ex, $"Failed to create instance of patch: {type.FullName}");
-          }
-        }
-
-        return _patches;
-      }
-    }
-
-    public static string GetBinSubDir() {
-      try {
-        return Utilities.GetConfigsPath();
-      }
-      catch (NullReferenceException) {
-        return "Win64_Shipping_Client";
-      }
     }
 
   }
