+++ src/CommunityPatch/Patches/MaxFoodVarietyPartyMoralePatch.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using System.Reflection;
-using System.Linq;
 using System.Runtime.CompilerServices;
 using TaleWorlds.CampaignSystem;
 using TaleWorlds.Core;
