+++ src/CommunityPatch/Patches/Perks/Steward/StewardSupremeAuthorityPatch.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using HarmonyLib;
